% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gis.R
\name{addHorizontalDistances}
\alias{addHorizontalDistances}
\title{Helper function: add horizontal distances from left boundary coordinates}
\usage{
addHorizontalDistances(gisData, leftBoundaryRow = 1)
}
\arguments{
\item{gisData}{data.frame as retrieved by importShapefiles()}

\item{leftBoundaryRow}{row number index of "gisData" which contains the left
model boundary (Default: 1)}
}
\value{
Model coordinates instead of "real" world coordinates
}
\description{
Helper function: add horizontal distances from left boundary coordinates
}
\examples{
shp.dir <- system.file("extdata", "qgis", package="kwb.demeau")
shp.files <- dir(path = shp.dir, pattern = ".shp", full.names = TRUE)
### Store GIS data in R data.frame "gisData"
gisData <- importShapefiles(shp.files)
### Add horizontal distances from left boundary and ignore all features that
### have larger distances than "right" boundary
gisData <- addHorizontalDistances(gisData)
### Plot horizontal model coordinates
maxVertical <- -abs(max(gisData$fcBottom, na.rm=TRUE))
}

