% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/monitoring.R
\name{calcWaterLevelChange}
\alias{calcWaterLevelChange}
\title{Helper function: calculate water level change}
\usage{
calcWaterLevelChange(refDate = "2009-03-02", df)
}
\arguments{
\item{refDate}{reference date for start (Default: "2009-03-02")}

\item{df}{data frame with structure like moniDat$agg$dailyMedian}
}
\description{
Helper function: calculate water level change
}

