% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrateModel}
\alias{calibrateModel}
\title{Calibration: single parameter}
\usage{
calibrateModel(preparedHeatModel, moniDat, obsPoints = "*",
  parameterName = "ratioKzKh", parameterRange = c(0.01, 0.1),
  objState = "waterLevelChange", objCrit = "RMSE", ...)
}
\arguments{
\item{preparedHeatModel}{as retrieved by prepareHeatModel()}

\item{moniDat}{moniDat}

\item{obsPoints}{regular expression of observation points/wells to be
included for goodness of fit calculation (Default: *, i.e. all); if only
BSV-6_3, then: "BSV-6-3"}

\item{parameterName}{parameterName (Default: "ratioKzKh")}

\item{parameterRange}{parameterRange (Default: c(0.01, 0.1))}

\item{objState}{model state variable to be optimised either "waterLevelChange"
or "temp" (for temperature) (Default: "waterLevelChange")}

\item{objCrit}{vector with performance parameters produced by function
hydroGOF:gof(), Default: "RMSE" (valid parameters: "ME", MAE", "MSE", "RMSE",
"NRMSE", "PBIAS", "RSR", "rSD", "NSE", "mNSE", "rNSE", "d", "md", "rd", "cp",
"r", "R2", "bR2", "KGE", "VE"), ATTENTION: currently optimising is implemented
to MINIMISE the value of only ONE selected objCrit parameter. Thus please make
sure that the best model fit results of the MINIMUM of the selected parameter}

\item{...}{further arguments passed to hydroGOF::gof()}
}
\value{
calibration results
}
\description{
Calibration: single parameter
}

