% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comparison.R
\name{compareModelledMeasured}
\alias{compareModelledMeasured}
\title{Compare measured & modelled resutls}
\usage{
compareModelledMeasured(heatModel, moniDat, toPlot = TRUE)
}
\arguments{
\item{heatModel}{object as retrieved by runHeatModel()}

\item{moniDat}{as retrieved by processingData()}

\item{toPlot}{If TRUE results are plotted (Default: TRUE)}
}
\value{
Plot/List of water level, water level change & temperature of measured vs.
modelled data
}
\description{
Compare measured & modelled resutls
}

