% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/monitoring.R
\name{convertToListAndAddMoniColumns}
\alias{convertToListAndAddMoniColumns}
\title{Monitoring: convert to list & add monitoring columns}
\usage{
convertToListAndAddMoniColumns(df, keyFields = "myDateTime")
}
\arguments{
\item{df}{dataframe containing the data to be transformed}

\item{keyFields}{keyFields to be used as keyFields in kwb.hsMatrixToListForm
(Default: "myDateTime")}
}
\description{
Monitoring: convert to list & add monitoring columns
}

