% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{convHeadBoundary}
\alias{convHeadBoundary}
\title{Helper function: convert head boundary}
\usage{
convHeadBoundary(leftHead = TRUE, modelStructure, depthToWaterTable,
  hydraulicGradient)
}
\arguments{
\item{leftHead}{if TRUE left head, if FALSE right head (Default: TRUE)}

\item{modelStructure}{modelStructure as retrieved by}

\item{depthToWaterTable}{depthToWaterTable}

\item{hydraulicGradient}{hydraulicGradient}
}
\value{
head boundary meta info
}
\description{
Helper function: convert head boundary
}
