% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conversion.R
\name{convModelCoordinatesToNodes}
\alias{convModelCoordinatesToNodes}
\title{Conversion: model coordinates to nodes}
\usage{
convModelCoordinatesToNodes(modelCoords, dx = 1, dy = 1)
}
\arguments{
\item{modelCoords}{as retrieved by convRealToModelCoordinates()}

\item{dx}{horizontal model grid spacing (Default: 1)}

\item{dy}{vertical model grid spacing (Default: 1)}
}
\value{
Model nodes corresponding to model coordinates
}
\description{
Conversion: model coordinates to nodes
}
\examples{
shp.dir <- system.file("extdata", "qgis", package="kwb.demeau")
shp.files <- dir(path = shp.dir, pattern = ".shp", full.names = TRUE)
gisData <- importShapefiles(shp.files)
### Optionally remove some features
# gisData <- removeFeatures(gisData = gisData, ignoreFeatureIDs = c(3,20))
modelCoords <- convRealToModelCoordinates(gisData, dx=1, dy=1)
modelNodes <- convModelCoordinatesToNodes(modelCoords=modelCoords)
}

