% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conversion.R
\name{convRealCoordinatesToNodes}
\alias{convRealCoordinatesToNodes}
\title{Conversion: "real" coordinates to model nodes}
\usage{
convRealCoordinatesToNodes(gisData, dx = 1, dy = 1, y = NULL)
}
\arguments{
\item{gisData}{data.frame() as retrieved by importShapefiles()}

\item{dx}{horizontal model grid spacing (Default: 1)}

\item{dy}{vertical model grid spacing (Default: 1)}

\item{y}{vertical model extent (Default: NULL, i.e. maximum filter screen depth below ground level)}
}
\value{
model nodes
}
\description{
By calling functions convRealToModelCoordinates() and  convModelCoordinatesToNodes()
}

