% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{convRealToModelCoordinates}
\alias{convRealToModelCoordinates}
\title{Conversion: "real" coordinates to model coordinates}
\usage{
convRealToModelCoordinates(gisData, dx = 1, dy = 1, y = NULL)
}
\arguments{
\item{gisData}{data.frame() as retrieved by importShapefiles()}

\item{dx}{horizontal model grid spacing (Default: 1)}

\item{dy}{vertical model grid spacing (Default: 1)}

\item{y}{vertical model extent (Default: NULL, i.e. maximum filter screen depth below ground level)}
}
\description{
Conversion: "real" coordinates to model coordinates
}
\examples{
shp.dir <- system.file("extdata", "qgis", package="kwb.demeau")
shp.files <- dir(path = shp.dir, pattern = ".shp", full.names = TRUE)
gisData <- importShapefiles(shp.files)
### Optionally remove some features 
# gisData <- removeFeatures(gisData = gisData, ignoreFeatureIDs = c(3,20))
modelCoords <- convRealToModelCoordinates(gisData, dx=1, dy=1)
xyplot(y ~ x, groups = Name, data=modelCoords, pch=16, auto.key=TRUE)
}
