% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{defineHeadBoundary}
\alias{defineHeadBoundary}
\title{Helper function: define head boundary}
\usage{
defineHeadBoundary(head, nly, temp, ntx = 4, ntc = 1)
}
\arguments{
\item{head}{as retrieved by convHeadBoundary()}

\item{nly}{number of vertical nodes}

\item{temp}{temperature at boundary}

\item{ntx}{vector with node type identifier for boundary conditions. 0 (for 
no specified boundary (needed for resetting some nodes after initial recharge 
period); 1 (for specified pressure head); 2 (for specified flux per unit 
horizontal surface area in units of L/T); 3 (for possible seepage face);
4 (for specified total head); 5 (for evaporation, Note: is not implemented
yet!); 6 (for specified volumetric flow in units of L3/T). 7 (for gravity 
drain). (The gravity drain boundary condition allows gravity driven vertical
flow out of the domain assuming a unit vertical hydraulic gradient. Flow 
into the domain cannot occur.)"}

\item{ntc}{vector with node type identifier for transport boundary conditions. 
0 (for no specified boundary); 1 (for specified temperatures), (Default: 1)}
}
\value{
boundary head parameterisation
}
\description{
Helper function: define head boundary
}
