% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominantTravelTime.R
\name{dominantTravelTimes}
\alias{dominantTravelTimes}
\title{Dominant travel time: data preprocessing}
\usage{
dominantTravelTimes(concModelled, offset = 0.01)
}
\arguments{
\item{concModelled}{as retrieved by kwb.demeau::soluteModelled()}

\item{offset}{offset (Default: 0.01) used for filtering soluteModel results
(i.e. maxConc/2 +- offset)}
}
\value{
list with dominant travel times with sublists "raw" (multiple values for 
each TIME_day possible) and "agg" (median "TIME_day" and "modelled" concentration)
}
\description{
Dominant travel time: data preprocessing
}
