% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dominantTravelTime.R
\name{dominantTravelTimesAgg}
\alias{dominantTravelTimesAgg}
\title{Helper function for dominantTravelTimes}
\usage{
dominantTravelTimesAgg(domTimes)
}
\arguments{
\item{domTimes}{intermediate result of function kwb.demeau::dominantTravelTimes}
}
\value{
list with aggregated dominant travel times (median!)
}
\description{
Helper function for dominantTravelTimes
}

