% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{downloadMeteoData}
\alias{downloadMeteoData}
\title{Monitoring: download meteo data}
\usage{
downloadMeteoData(startYear = 2008, endYear = 2014)
}
\arguments{
\item{startYear}{1.1.startYear (Default: 1.1.2008)}

\item{endYear}{31.12.endYear (Default: 31.12.2014)}
}
\description{
Monitoring: download meteo data
}
