% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{filterMoniData}
\alias{filterMoniData}
\title{Helper function: filter monitoring data}
\usage{
filterMoniData(locations = NULL, locationsCol = "moniLocation",
  paras = NULL, paraCol = "moniParName", minDate = "2009-03-02",
  maxDate = "2009-04-06", df)
}
\arguments{
\item{locations}{(Default: NULL)}

\item{locationsCol}{(Default: "moniLocation")}

\item{paras}{(Default: NULL)}

\item{paraCol}{(Default: "moniParName")}

\item{minDate}{(Default: "2009-03-02")}

\item{maxDate}{(Default: "2009-04-06")}

\item{df}{data.frame structure like: moniDat$agg$dailyMedian}
}
\description{
Helper function: filter monitoring data
}
