% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{fitnessAdaptedModelConfiguration}
\alias{fitnessAdaptedModelConfiguration}
\title{Calibration: helper function "fitnessAdaptedModelConfiguration" ( called by 
function calibrateModel()}
\usage{
fitnessAdaptedModelConfiguration(parameterValue, parameterName,
  preparedHeatModel, objState = "waterLevelChange", objCrit = "RMSE",
  moniDat, obsPoints, ...)
}
\arguments{
\item{parameterValue}{parameterValue}

\item{parameterName}{parameterName}

\item{preparedHeatModel}{as retrieved by prepareHeatModel()}

\item{objState}{model state variable to be optimised either "waterLevelChange"
or "temp" (for temperature) (Default: "waterLevelChange")}

\item{objCrit}{vector with performance parameters produced by function 
hydroGOF:gof(), Default: "RMSE" (valid parameters: "ME", MAE", "MSE", "RMSE", 
"NRMSE", "PBIAS", "RSR", "rSD", "NSE", "mNSE", "rNSE", "d", "md", "rd", "cp", 
"r", "R2", "bR2", "KGE", "VE")}

\item{moniDat}{moniDat}

\item{obsPoints}{regular expression of observation points/wells to be 
included for goodness of fit calculation (Default: *, i.e. all); if only 
BSV-6_3, then: "BSV-6-3"}

\item{...}{further arguments passed to hydroGOF::gof()}
}
\value{
fitness of model configuration
}
\description{
Calibration: helper function "fitnessAdaptedModelConfiguration" ( called by 
function calibrateModel()
}
