% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comparison.R
\name{fitnessWithLabel}
\alias{fitnessWithLabel}
\title{Compare measured & modelled results}
\usage{
fitnessWithLabel(heatModel, moniDat, objState = "waterLevelChange",
  objCrit = c("RMSE", "R2"), plotIt = TRUE, plot.type = "b",
  cex.label = 1, main = NULL, performance.in.label = TRUE, ...)
}
\arguments{
\item{heatModel}{object as retrieved by runHeatModel()}

\item{moniDat}{as retrieved by processingData()}

\item{objState}{model state variable to be optimised either "waterLevelChange"
or "temp" (for temperature) (Default: "waterLevelChange")}

\item{objCrit}{vector with performance parameters produced by function
hydroGOF:gof(), Default: "RMSE" (valid parameters: "ME", MAE", "MSE", "RMSE",
"NRMSE", "PBIAS", "RSR", "rSD", "NSE", "mNSE", "rNSE", "d", "md", "rd", "cp",
"r", "R2", "bR2", "KGE", "VE")}

\item{plotIt}{if TRUE lattice plot will be produced (Default: TRUE)}

\item{...}{additional parameters passe to plot function lattice:xyplot()}
}
\value{
Plot of water level, water level change & temperature of measured vs.
modelled data
}
\description{
Compare measured & modelled results
}

