% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genuchtenModel.R
\name{genuchten}
\alias{genuchten}
\title{vanGenuchten : helper for multiple models}
\usage{
genuchten(pressureHeads = -rev(seq(0, 1000, 5)), alpha = 1, beta = 2)
}
\arguments{
\item{pressureHeads}{pressureHeads}

\item{alpha}{alpha to be used for van Genuchten model}

\item{beta}{betaa to be used for van Genuchten model}
}
\value{
data.frame with columns "pressureHead, alpha, beta, effSaturation, Kr"
}
\description{
vanGenuchten : helper for multiple models
}
\references{
http://wwwbrr.cr.usgs.gov/projects/GW_Unsat/vs2di/hlp/solute/vanGenuchten.html
}

