% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genuchtenModel.R
\name{genuchtenModel}
\alias{genuchtenModel}
\title{vanGenuchten : single model}
\usage{
genuchtenModel(pressureHead, alpha, beta)
}
\arguments{
\item{pressureHead}{pressureHead}

\item{alpha}{alpha}

\item{beta}{beta}
}
\value{
data.frame with columns "pressureHead, alpha, beta, effSaturation, Kr"
}
\description{
vanGenuchten : single model
}
\references{
http://wwwbrr.cr.usgs.gov/projects/GW_Unsat/vs2di/hlp/solute/vanGenuchten.html
}
