% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genuchtenModel.R
\name{genuchtenModels}
\alias{genuchtenModels}
\title{vanGenuchten : multiple models}
\usage{
genuchtenModels(pressureHeads = -rev(seq(0, 6, 0.5)), alphas = seq(1, 2,
  0.5), betas = 1:5)
}
\arguments{
\item{pressureHeads}{pressureHeads}

\item{alphas}{vector of alphas to be used for multiple model construction}

\item{betas}{vector of betas to be used for multiple model construction}
}
\value{
data.frame with columns "pressureHead, alpha, beta, effSaturation, Kr"
}
\description{
vanGenuchten : multiple models
}
\references{
http://wwwbrr.cr.usgs.gov/projects/GW_Unsat/vs2di/hlp/solute/vanGenuchten.html
}
