% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getFeatures}
\alias{getFeatures}
\title{Helper function: Gets feature information}
\usage{
getFeatures(gisData, addColNames = NULL)
}
\arguments{
\item{gisData}{data.frame with gis features as retrieved by importShapefile(s)()}

\item{addColNames}{vector with additional colnames for output data.frame (Default: NULL),
for valid inputs check: colNames(gisData)}
}
\value{
Return the features with attributes feature id (fID), shapefile name 
(shape.name) and feature name (Name)
}
\description{
Aggregates gis features by shape.name & id add adds new column fID and 
returns additional metainformation (parameter )
}
