% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{importData}
\alias{importData}
\title{Monitoring: excel data import}
\usage{
importData(xlsPath = NULL, metaTables = c("metadataTable", "commentTable"),
  moniTables = c("bsv4", "bsvAll", "tuberia", "inflow"),
  ignoredParNames = c("Temp_graphics_C", "DiverBaro_graphFictive_cm",
  "DiverBaro_cm"))
}
\arguments{
\item{xlsPath}{full path to excel file http://192.168.22.12/svn/kwb/DEMEAU/Work 
Areas/WA1 MAR/TracerSVH/datosTOT_StVicen+_jun2008-abr2009OK.xls", if FALSE 
already imported data object moniDat is loaded (Default: NULL)}

\item{metaTables}{vector with names of tables with meta information 
(Default: "metadataTable", "commentTable")}

\item{moniTables}{vector with names of tables with monitoring data to be 
imported (Default: "bsv4", "bsvAll", "tuberia", "inflow")}

\item{ignoredParNames}{vector of parNames that are ignored for storing 
(Default:Temp_graphics_C, DiverBaro_graphFictive_cm,  DiverBaro_cm: syntetical 
or calculated parameters!)}
}
\description{
Monitoring: excel data import
}
\examples{
### xlsDir needs to be set correctly !!!!!
xlsDir <- "C:/Users/mrustl/Documents/WC_Server/DEMEAU/Work Areas/WA1 MAR/TracerSVH"
xlsFile <- "datosTOT_StVicen_jun2008-abr2009OK.xls"
xlsPath <- file.path(xlsDir, xlsFile)
##importData(xlsPath=xlsPath)

#### Loading with stored moniDat.RData object
#importData()
}
