% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{importShapefile}
\alias{importShapefile}
\title{GIS: imports shapefiles & dbf}
\usage{
importShapefile(shp.path)
}
\arguments{
\item{shp.path}{full path to file (with or without file extension ".shp")}
}
\value{
Imported GIS shapefile as R data.frame
}
\description{
GIS: imports shapefiles & dbf
}
