% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gis.R
\name{importShapefiles}
\alias{importShapefiles}
\title{GIS: imports shapefiles & dbf}
\usage{
importShapefiles(shp.files = dir(path = system.file("extdata", "qgis", package
  = "kwb.demeau"), pattern = ".shp", full.names = TRUE))
}
\arguments{
\item{shp.files}{vector with full paths to shapefiles "Boundary", "Ponds",
"Observation wells" to be imported (with or without file extension ".shp")}
}
\value{
Imported GIS shapefiles in an R data.frame
}
\description{
GIS: imports shapefiles & dbf
}
\examples{
shp.dir <- system.file("extdata", "qgis", package="kwb.demeau")
shp.files <- dir(path = shp.dir, pattern = ".shp", full.names = TRUE)
### Store GIS data in R data.frame "gisData"
gisData <- importShapefiles(shp.files)
### Plot imported GIS data:
}

