% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gis.R
\name{leftBoundaryInPolygon}
\alias{leftBoundaryInPolygon}
\title{Helper function: check whether left model boundary lies within polygon}
\usage{
leftBoundaryInPolygon(gisData, leftBoundaryRow = 1)
}
\arguments{
\item{gisData}{data.frame as retrieved by importShapefiles()}

\item{leftBoundaryRow}{row number index of "gisData" which contains the left}
}
\description{
Helper function: check whether left model boundary lies within polygon
}

