% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{listToMatrixForm}
\alias{listToMatrixForm}
\title{Monitoring: convert data to matrix forma}
\usage{
listToMatrixForm(df)
}
\arguments{
\item{df}{data frame with structure like: moniDat$agg$dailyMedian}
}
\description{
Monitoring: convert data to matrix forma
}
