% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{mainLabel}
\alias{mainLabel}
\title{Helper function for main label of comparision plot}
\usage{
mainLabel(model)
}
\arguments{
\item{model}{object as retrieved by runHeatModel()}
}
\value{
main label
}
\description{
Helper function for main label of comparision plot
}
