% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{modelConfiguration}
\alias{modelConfiguration}
\title{Model: prepare model configuration}
\usage{
modelConfiguration(modelStructure, pondTemp = 20, gwTempIni = 12,
  infRate = 0.03, depthToWaterTable = 6, hydraulicGradient = 0.001,
  bnd = list(tmp = gwTempIni, ntx = 4, ntc = 1),
  hk = kwb.vs2dh::vs2dh.ConfigureGenuchten(ratioKzKh = 1, ss = 0, satKh =
  750, porosity = 0.2, alpha = 2.3, rmc = 0, beta = 5.8),
  ht = kwb.vs2dh::vs2dh.ConfigureTrans(), iniOutputTime = 1/(3600 *
  24), minSimTime = 0.5, maxSimTime = 31, outputTimeStep = 1,
  solver = kwb.vs2dh::vs2dh.ConfigureBasicSolver(),
  rSolver = kwb.vs2dh::vs2dh.ConfigureRechargePeriodSolver())
}
\arguments{
\item{modelStructure}{as retrieved by convRealCoordinatesToNodes()}

\item{pondTemp}{constant pond temperature (default: 20)}

\item{gwTempIni}{initial groundwater temperature (default: 12)}

\item{infRate}{infiltration rate per unit area (Default: 0.9514151 m/d)}

\item{depthToWaterTable}{water table below ground level (default: 6 m)}

\item{hydraulicGradient}{hydraulic gradient between left & right model boundary 
(Default: 0), if positive flow is from left to right, if negative from right 
to left}

\item{bnd}{list of structure list(temp=VALUE, ntx=VALUE, ntx=VALUE) passed to 
function defineHeadBoundary (i.e. boundary with seepage face), if bnd=NULL 
left/right boundaries are no-flow boundaries}

\item{hk}{hydraulic properties of soil as retrieved by kwb.vs2dh::vs2dh.ConfigureGenuchten()}

\item{ht}{transport properties of soil as retrieved by kwb.vs2dh::vs2dh.ConfigureTrans()}

\item{iniOutputTime}{automatically output results after 1 second of simulation}

\item{minSimTime}{minimum simulation time in days (default: 0.5)}

\item{maxSimTime}{maximum simulation time in days (default: 31)}

\item{outputTimeStep}{at which timestep are the results printed (default: 1 ), i.e. each day}

\item{solver}{general solver (Default: kwb.vs2dh::vs2dh.ConfigureBasicSolver())}

\item{rSolver}{recharge period solver (Default: kwb.vs2dh::vs2dh.ConfigureRechargePeriodSolver())}
}
\value{
SVH model configuration
}
\description{
Model: prepare model configuration
}
\examples{
\dontrun{
### Importing GIS features
shp.dir <- system.file("extdata", "qgis", package="kwb.demeau")
shp.files <- dir(path = shp.dir, pattern = ".shp", full.names = TRUE)
gisData <- importShapefiles(shp.files)

### Optionally remove some features 
gisData <- removeFeatures(gisData = gisData, ignoreFeatureIDs = c(3,20))
modelStructure <- convRealCoordinatesToNodes(gisData = gisData)

### Model config
conf <- modelConfiguration(modelStructure = modelStructure)
               
### Running the configuration in VS2DH
res <- kwb.vs2dh::vs2di.runConfig(conf = conf, openTargetDir = TRUE)

### Plotting results
kwb.vs2dh::vs2dh.plotObservationPoints(
  paras = "TEMP", paraLabel = "Temperature", data = res$obsPoints
)
kwb.vs2dh::vs2dh.plotVariables(para = "Temp", data = res$variables)
}

}
\seealso{
\code{defineHeadBoundary} for valid additional arguments
}
