% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{modelCoordinatesToFeature}
\alias{modelCoordinatesToFeature}
\title{Conversion: feature parameterisation}
\usage{
modelCoordinatesToFeature(x = c(100, 220), y = 0, dx = 0.5,
  dy = 0.5, pondDepth = verticalCmToMeters(cm = 1.2), steepness = 45)
}
\arguments{
\item{x}{absolute horizontal min/max coordinates of model feature. L, Default: c(100,220)}

\item{y}{absolute vertical min/max coordinates of model feature. L, Default: c(0,0)}

\item{dx}{dx}

\item{dy}{dy}

\item{pondDepth}{Depth of pond below ground level (Default: verticalCmToMeters(cm = 1.2), 
i.e. 3.4m, derived from Figure 8, p.11 ENSAT report)}

\item{steepness}{of pond bank (in degree) only used for ponds if "pondDepth" 
is defined (Default: 45 degree)}
}
\value{
Parameterisation of model features
}
\description{
Conversion: feature parameterisation
}
