% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calibrate.R
\name{modelFitness}
\alias{modelFitness}
\title{Calibration: helper function "modelFitness" (called by function modelFitnessAggregated())}
\usage{
modelFitness(modelledMeasured, obsPoints = "*", ...)
}
\arguments{
\item{modelledMeasured}{as retrieved by compareModelledMeasured() either $temp or
$waterLevelChange}

\item{obsPoints}{regular expression of observation points/wells to be
included for goodness of fit calculation (Default: *, i.e. all); if only
BSV-6_3, then: "BSV-6-3"}

\item{...}{further arguments passed to hydroGOF::gof()}
}
\value{
matrix with goodness of fit criteria
}
\description{
Calibration: helper function "modelFitness" (called by function modelFitnessAggregated())
}

