% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper.R
\name{nodeIdToRowColumn}
\alias{nodeIdToRowColumn}
\title{Helper function: converting node ID to row/column}
\usage{
nodeIdToRowColumn(nodeID, numberOfRows)
}
\arguments{
\item{nodeID}{vector of node ids}

\item{numberOfRows}{number of rows (here: nly)}
}
\value{
row (yNode) and column of node ID in grid
}
\description{
Helper function: converting node ID to row/column
}

