% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comparison.R
\name{orderedDataFrame}
\alias{orderedDataFrame}
\title{Helper function to order data frame by two columns (col1, col2)}
\usage{
orderedDataFrame(df, col1 = "Name", col2 = "TIME_day")
}
\arguments{
\item{df}{data.frame}

\item{col1}{first column to be used for ordering (Default: Name)}

\item{col2}{second column to be used for ordering (Default: TIME_day)}

\item{...}{further arguments passed to function order(), e.g. decreasing = TRUE}
}
\value{
ordered (default: accending) data.frame according to col1 & col2
}
\description{
Helper function to order data frame by two columns (col1, col2)
}

