% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genuchtenModel.R
\name{plotGenuchtenModels}
\alias{plotGenuchtenModels}
\title{vanGenuchten : plot models}
\usage{
plotGenuchtenModels(models = genuchtenModels(), ...)
}
\arguments{
\item{models}{as retrieved by genuchtenModels()}

\item{...}{further arguments passed to lattice::xyplot()}
}
\value{
plot of all genuchten models
}
\description{
vanGenuchten : plot models
}
