% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotModelStructure}
\alias{plotModelStructure}
\title{Plot model Structure}
\usage{
plotModelStructure(df, xColName = "x", yColName = "y")
}
\arguments{
\item{df}{data.frame as retrieved by convRealToModelCoordinates() or
convModelCoordinatesToNodes()}

\item{xColName}{name of x column to plot}

\item{yColName}{name of y column to plot}
}
\description{
Plot model Structure
}
