% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMonitoringData}
\alias{plotMonitoringData}
\title{Monitoring: plot data}
\usage{
plotMonitoringData(moniParName = "WaterLevel_cm",
  moniLocationPattern = "BSV|Tuberia", minDate = "2000-01-01",
  maxDate = "2015-01-01", groups = FALSE, df)
}
\arguments{
\item{moniParName}{(Default: "WaterLevel_cm")}

\item{moniLocationPattern}{(Default: "BSV|Tuberia")}

\item{minDate}{(Default: "2000-01-01")}

\item{maxDate}{(Default: "2015-01-01")}

\item{groups}{(Default: FALSE)}

\item{df}{data frame}
}
\description{
Monitoring: plot data
}
