% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMonitoringWithTwoYAxes}
\alias{plotMonitoringWithTwoYAxes}
\title{Monitoring: plot data with two y axex}
\usage{
plotMonitoringWithTwoYAxes(parY1 = "WaterLevelChange_cm", parY2 = "Temp_C",
  labelParY1 = "Water level change (cm)", labelParY2 = "Temperature (C)",
  df, ...)
}
\arguments{
\item{parY1}{name of parameter for Y1 axis (see: colnames(listToMatrixForm(df)),
(Default: "WaterLevelChange_cm")}

\item{parY2}{name of parameter for Y2 axis (see: colnames(listToMatrixForm(df)),
(Default: "Temp_C")}

\item{labelParY1}{user defined label for Y1 for legend (Default: 
"Water level change (cm)")}

\item{labelParY2}{user defined label for Y2 for legend (Default: 
"Temperature (degree C)")}

\item{df}{data.frame with structure like moniDat$agg$dailyMedian}

\item{...}{further parameters passed to xyplot()}
}
\description{
Monitoring: plot data with two y axex
}
