% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/processingData.R
\name{processingData}
\alias{processingData}
\title{Processing: data preprocessing}
\usage{
processingData(rawData)
}
\arguments{
\item{rawData}{raw data imported with importData()}
}
\description{
Processing: data preprocessing
}

