% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper.R
\name{removeFeatures}
\alias{removeFeatures}
\title{Helper function: remove features which should be ignored}
\usage{
removeFeatures(gisData, ignoreFeatureIDs = NULL)
}
\arguments{
\item{gisData}{data.frame with gis features as retrieved by importShapefile(s)()}

\item{ignoreFeatureIDs}{should be a valid feature id column gisData$fID,
for possible values check: unique(gisData$fID)}
}
\value{
input gisData (possibly removed by ignoredFeatureIDs)
}
\description{
Helper function: remove features which should be ignored
}

