% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{removeFileExtension}
\alias{removeFileExtension}
\title{Helper function: removes file extensions (copied from "kwb.quantum" package)}
\usage{
removeFileExtension(x)
}
\arguments{
\item{x}{file.path}
}
\value{
file path without extension (i.e. without ".*")
}
\description{
Helper function: removes file extensions (copied from "kwb.quantum" package)
}
