% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{renameValues}
\alias{renameValues}
\title{Helper function: rename values}
\usage{
renameValues(df, colName, oldVal, newVal)
}
\arguments{
\item{df}{data.frame to be modified}

\item{colName}{olName containing the values to be changed}

\item{oldVal}{old value}

\item{newVal}{new value}
}
\description{
Helper function: rename values
}
