% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model.R
\name{runHeatModel}
\alias{runHeatModel}
\title{Run heat model: wrapper for vs2di.runConfig()}
\usage{
runHeatModel(preparedHeatModel, tDir = tempdir(), openTargetDir = FALSE,
  ...)
}
\arguments{
\item{preparedHeatModel}{prepared heat model as retrieved by  prepareModel(type="heat")}

\item{tDir}{target directory where vs2dh model input/output files should be
stored. (Default: tempdir())}

\item{openTargetDir}{If TRUE: model directory with heat model results will be
opened in Windows explorer (Default: FALSE)}

\item{...}{additional arguements passed to function kwb.vs2dh::vs2di.runConfig()}
}
\value{
List with heat model results ($res) and configuration
}
\description{
Run heat model: wrapper for vs2di.runConfig()
}

