% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{runSoluteModel}
\alias{runSoluteModel}
\title{Run solute transport model: wrapper for vs2di.runConfig()}
\usage{
runSoluteModel(preparedSoluteModel, tDir = tempdir(),
  returnOutput = TRUE, openTargetDir = FALSE, ...)
}
\arguments{
\item{preparedSoluteModel}{prepared solute transport model as retrieved by  prepareModel(type="solu")}

\item{tDir}{target directory where vs2dh model input/output files should be 
stored. (Default: tempdir())}

\item{returnOutput}{Default: TRUE}

\item{openTargetDir}{If TRUE: model directory with heat model results will be
opened in Windows explorer (Default: FALSE)}

\item{...}{additional arguements passed to function kwb.vs2dh::vs2di.runConfig()}
}
\value{
List with heat model results ($res) and configuration
}
\description{
Run solute transport model: wrapper for vs2di.runConfig()
}
