% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{selectModelled}
\alias{selectModelled}
\title{Compare measured & modelled results}
\usage{
selectModelled(modDf, parName = "TEMP", func = median)
}
\arguments{
\item{modDf}{modDf}

\item{parName}{(Default: "TEMP")}

\item{func}{(Default: median)}
}
\value{
Plot of water level, water level change & temperature of measured vs.
modelled data
}
\description{
Compare measured & modelled results
}
