% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper.R
\name{setSoilZero}
\alias{setSoilZero}
\title{Helper function: set soil ITEX to zero}
\usage{
setSoilZero(nodes, soilGrid, shape.name = "Ponds")
}
\arguments{
\item{nodes}{as retrieved by convModelCoordinatesToNodes() or
convRealCoordinatesToNodes()}

\item{soilGrid}{soil grid (as retrieved by kwb.vs2dh::vs2dh.ConfigureSoilGrid())}

\item{shape.name}{name of polygon shape files with ponds (Default: "Ponds)}
}
\value{
soil grid (with possible ITEX values set to zero)
}
\description{
Helper function: set soil ITEX to zero
}

