% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comparison.R
\name{soluteModelled}
\alias{soluteModelled}
\title{Compare measured & modelled resutls}
\usage{
soluteModelled(soluteModel, offset = 0.01, toPlot = TRUE, ...)
}
\arguments{
\item{soluteModel}{object as retrieved by runSoluteModel()}

\item{offset}{offset (Default: 0.01) used for filtering soluteModel results
(i.e. maxConc/2 +- offset)}

\item{toPlot}{If TRUE results are plotted (Default: TRUE)}

\item{...}{additional parameters passe to plot function lattice:xyplot()}
}
\value{
Plot/List of water level, water level change & temperature of measured vs.
modelled data
}
\description{
Compare measured & modelled resutls
}

