% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper.R
\name{verticalCmToMeters}
\alias{verticalCmToMeters}
\title{Helper function: converting vertical scale into meters (Figure. 08, page 11, ENSAT report)}
\usage{
verticalCmToMeters(cm)
}
\arguments{
\item{cm}{measured centimeters in Figure 8}
}
\value{
vertical length in meters
}
\description{
Helper function: converting vertical scale into meters (Figure. 08, page 11, ENSAT report)
}

